Reorganize files into product package
